+++ apps/opik-frontend/src/components/layout/SideBar/SideBar.tsx
@@ -290,6 +290,7 @@ const SideBar: React.FunctionComponent<SideBarProps> = ({
     {
       placeholderData: keepPreviousData,
       refetchInterval: RUNNING_OPTIMIZATION_REFETCH_INTERVAL,
+      enabled: !!workspaceName,
     },
   );
